requirements.txt
python-dotenv==1.1.1
streamlit==1.50.0
langchain==0.3.27
langchain-groq==0.3.8
langchain-chroma==0.2.6
langchain-community==0.3.31
unstructured
tiktoken
langchain-chroma
langchain-huggingface==0.3.1
fastuuid==0.13.5
shutil
